% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abb_redcap_events.R
\name{abb_redcap_events}
\alias{abb_redcap_events}
\title{Abbreviate REDCap event names}
\usage{
abb_redcap_events(x, long_short = "long")
}
\arguments{
\item{x}{A dataframe containing variable "redcap_event_name".}

\item{long_short}{A character 'long' (default) or 'short' indicating whether
to use long or short variable abbreviations.}
}
\value{
A dataframe with new variable "redcap_event".
}
\description{
shorten "redcap_event_name" values for easier coding.
}
\examples{
(x <- data.frame(cp_ptid = c("50-22-0001-M1", "50-22-0001-B1"),
redcap_event_name = c("Enrollment (Arm 1: Mothers)",
                     "Vaccination & PMTCT (Arm 2: Babies)")))

abb_redcap_events(x)
abb_redcap_events(x, long_short = 'short')
}
