% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_date.R
\name{abc_date}
\alias{abc_date}
\title{Make an ABC-style date}
\usage{
abc_date(y = NULL, m = NULL, d = NULL, ymd = NULL)
}
\arguments{
\item{y}{A year. Character or numeric, 2 or 4 digits.}

\item{m}{A month. Character (full name, abbreviation, or character numbers) or numeric (2 or 4 digits).}

\item{d}{A date. Character or numeric (2 or 4 digits or characters).}

\item{ymd}{A character string with format YYYY-mm-dd.}
}
\value{
A character string with format YYYY_mmbbb_dd.
}
\description{
Creates a string following the ABC date convention to prevent any ambiguity.
For example: "2024_02feb_08." Accepts a variety of inputs and formats. If
\code{abc_date()} is used without any input arguments, or is missing year (y),
month (m), or date (d) arguments, today's date will be output (using
\code{Sys.Date()}).
}
\examples{
abc_date()
abc_date(y = '2023', m = '5', d = 6)
abc_date(y = 23, m = 'April', d = 06)
abc_date(ymd = "2024-01-20")
}
