\name{abctools-package}
\alias{abctools-package}
\alias{abctools}
\docType{package}
\title{Routines for ABC inference, including methods for summary statistics selection. }
\description{This package contains functions implementing recent methods for selecting summary statistics in ABC inference.}
\details{
\tabular{ll}{
Package: \tab abctools\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 10-02-2013\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{Matt Nunes and Dennis Prangle

Maintainer: Matthew Nunes <m.nunes@lancaster.ac.uk>
}
\references{
For details on methods for summary statistics selection, see\cr

Blum, M. G. B, Nunes, M. A., Prangle, D. and Sisson, S. A. (2013) A comparative review of dimension reduction methods in approximate Bayesian computation. \emph{Stat. Sci. (to appear)}.\cr\cr
Fearnhead, P. and Prangle, D. (2012) Constructing summary statistics for approximate Bayesian computation: semi-automatic approximate Bayesian computation. \emph{J. R. Stat. Soc. B} \bold{74}, Part 3, 1--28.\cr\cr
Joyce, P. and P. Marjoram (2008) Approximately sufficient statistics and Bayesian computation. \emph{Stat. Appl. Gen. Mol. Biol.} \bold{7} Article 26.\cr\cr
Nunes, M. A. and Balding, D. J. (2010) On Optimal Selection of Summary Statistics for Approximate Bayesian Computation.  \emph{Stat. Appl. Gen. Mol. Biol.} \bold{9}, Iss. 1, Art. 34. \cr\cr
Wegmann, D. et al. (2009) Efficient approximate Bayesian computation coupled With Markov chain Monte Carlo Without Likelihood. \emph{Genetics} \bold{182} (4), 1207--1218.\cr\cr

For extra scripts used for \code{pls.abc}, see:

Wegmann, D. et al. (2010) ABCtoolbox: A versatile toolkit for approximate Bayesian computations. \emph{BMC Bioinf.} \bold{11}, 116--123.
}
\keyword{ package }
\keyword{methods}

