% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_to_freqfactor.R
\name{char_to_freqfactor}
\alias{char_to_freqfactor}
\title{Convert character variables to frequency factors}
\usage{
char_to_freqfactor(x, factor_levels)
}
\arguments{
\item{x}{A data frame.}

\item{factor_levels}{A (character) vector of frequency factors.}
}
\value{
A dataframe.
}
\description{
Convert character variables in a dataframe to frequency factors with given a
defined set of values.
}
\examples{
# example data
(x <- data.frame(var1 = 1:3,
                 var2 = c("Rarely", "Daily", NA),
                 var3 = rep(NA, 3)))

# original classes
sapply(x, class)

# specify frequency factor categories
(ff <- c('Not at all', 'Rarely', 'Some days', 'Most days', 'Daily'))

# use function
(x <- char_to_freqfactor(x, ff))

# new classes
sapply(x, class)
}
