% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_to_numeric.R
\name{char_to_numeric}
\alias{char_to_numeric}
\title{Character to numeric}
\usage{
char_to_numeric(x)
}
\arguments{
\item{x}{A dataframe.}
}
\value{
A dataframe.
}
\description{
change variables interpreted as character variables back to numeric.
}
\examples{
(x <- data.frame(var1 = c("a", "b", "c"),
                 var2 = c("01", "2,000", 3),
                 var3 = c(1, 2, 3)))

# leading zeros and special characters make numbers interpreted as characters
sapply(x, class)

x <- char_to_numeric(x)

# all variables with numbers are numeric
sapply(x, class)
}
