% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_vars_all_na.R
\name{drop_vars_all_na}
\alias{drop_vars_all_na}
\title{Drop columns or variables that are all NA}
\usage{
drop_vars_all_na(x)
}
\arguments{
\item{x}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Drop columns or variables that are all NA
}
\examples{
(x <- data.frame(a = 1:4, b = rep(NA, 4), c = 5:8, d = c(NA, 'B', 'C', 'D')))
drop_vars_all_na(x)
}
