% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets_exposed.R
\name{ets_exposed}
\alias{ets_exposed}
\title{ETS exposed}
\usage{
ets_exposed(x, freq_factor)
}
\arguments{
\item{x}{A dataframe containing lbt_closetosmokers, en_peoplesmoking,
en_ind_cig, en_ind_marig, en_wrk_cig, and en_wrk_marij variables.}

\item{freq_factor}{A named list containing elements "yes" and "no", which
contain character vectors to define categories. For example:
list(no = c('Not at all', 'Rarely'), yes = c('Some days', 'Most days', 'Daily'))}
}
\value{
A dataframe containing the new categorical variable
"calc_ets_exposed" with values: No, Yes and Unknown.
}
\description{
Environmental tobacco smoke exposed.
}
\examples{
# make frequency factor list
(smoke_factor <- list(no = c('Not at all', 'Rarely'),
                      yes = c('Some days', 'Most days', 'Daily')))

# example data
(x <- data.frame(cp_ptid = paste0("50-22-000", 1:3, "-M1"),
                 lbt_closetosmokers = c("I don't know", "No", "Yes"),
                 en_peoplesmoking = c(0, 0, 1),
                 en_ind_cig = c(rep('Not at all', 3)),
                 en_ind_marig = c("Not at all", "Not at all", 'Some days'),
                 en_wrk_cig = c("Sometimes", "Not at all", "Not at all"),
                 en_wrk_marij = c(rep('Not at all', 3))) )

# assess ETS exposure
ets_exposed(x, smoke_factor)
}
