% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_ptid.R
\name{fix_ptid}
\alias{fix_ptid}
\title{Fix ID variables}
\usage{
fix_ptid(x)
}
\arguments{
\item{x}{A dataframe containing variables with names "ptid."}
}
\value{
A dataframe.
}
\description{
Fix incorrectly entered participant ID variables.
}
\examples{
(x <- data.frame(cp_ptid = c("50-22-0001-M1", "50-22-0002-M1", "50-22-0003-M1"),
gps_ptid = c("001", "0002", "50-22-0003-M1"),
con_ptid = c("01", 002, "0003") ))

fix_ptid(x)
}
