% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_redcap_data.R
\name{get_redcap_data}
\alias{get_redcap_data}
\title{Get ABC data from REDCap with API token}
\usage{
get_redcap_data(token = NULL, token_path = NULL, ...)
}
\arguments{
\item{token}{A character vector containing the API token.}

\item{token_path}{A file path to the .R file containing the API token.}

\item{...}{optional arguments passed to \code{REDCapR::redcap_read_oneshot()}.}
}
\value{
A dataframe (tibble) containing ABC data from REDCap.
}
\description{
A wrapper function for \code{REDCapR::redcap_read_oneshot()} that simplifies
obtaining ABC data from REDCap.
}
\examples{
# path to .R script containing ABC API token
tp <- file.path(Sys.getenv("R_USER"), "R", "access_tokens", "abc_token.R")

# collect REDCap data
get_redcap_data(token_path = tp)
}
