% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_all_na_vars.R
\name{id_all_na_vars}
\alias{id_all_na_vars}
\title{Identify NA variables}
\usage{
id_all_na_vars(x)
}
\arguments{
\item{x}{A dataframe.}
}
\value{
A character vector consisting of a subset of column names of \code{x}.
}
\description{
Identify variables or columns that are all NA.
}
\examples{
(x<- data.frame(var1 = c(1, 2, 3),
                var2 = c(NA, NA, NA),
                var3 = c(3, NA, 4),
                var4 = c(NA, NA, NA) ) )
id_all_na_vars(x)
}
