% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_freq_factor_vars.R
\name{id_freq_factor_vars}
\alias{id_freq_factor_vars}
\title{ID frequency factor variables}
\usage{
id_freq_factor_vars(x, factor_levels)
}
\arguments{
\item{x}{A dataframe.}

\item{factor_levels}{A character vector containing frequency factors.}
}
\value{
A character vector of variable names.
}
\description{
Identify variables that contain values specified in a character vector.
}
\examples{
(x <- data.frame(var1 = c("Daily", "Rarely"), var2 = 1:2, var3 = c(NA, "Most days")))
id_freq_factor_vars(
    x = x,
    factor_levels = c('Not at all', 'Rarely', 'Some days', 'Most days', 'Daily')
    )
}
