% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_yesno_vars.R
\name{id_yesno_vars}
\alias{id_yesno_vars}
\title{ID "yes/no" variables}
\usage{
id_yesno_vars(x)
}
\arguments{
\item{x}{A dataframe.}
}
\value{
A character vector of variable names.
}
\description{
Identify yes/no variables. Recognizes a variety of values, e.g.: Yes, yes,
YES, Y, y.

Helpful on its own, also used in \code{yesno_to_logical()} function.
}
\examples{
(x <- data.frame(var1 = c('Yes', 'yes', 'no', 'No'), var2 = 1:4, var3 = c('yes', rep(NA, 3))))
id_yesno_vars(x)
}
