% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_returns.R
\name{remove_returns}
\alias{remove_returns}
\title{Remove returns}
\usage{
remove_returns(x)
}
\arguments{
\item{x}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Searches through variables of a dataframe and removes carriage returns in
free text.
}
\examples{
(x <- data.frame(var1 = "line1\r\nline2", var2 = 2) )
remove_returns(x)
}
