% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrk_ets_exposed.R
\name{wrk_ets_exposed}
\alias{wrk_ets_exposed}
\title{Work ETS exposed}
\usage{
wrk_ets_exposed(x, freq_factor)
}
\arguments{
\item{x}{A dataframe containing variables: en_wrk_cig and en_wrk_marij.}

\item{freq_factor}{A named list containing elements "yes" and "no", which
contain character vectors to define categories. For example:
list(no = c('Not at all', 'Rarely'), yes = c('Some days', 'Most days', 'Daily'))}
}
\value{
A dataframe containing the new logical variable
"calc_wrk_ets_exposed."
}
\description{
Work/occupational environmental tobacco smoke exposed.
}
\examples{
#' # make frequency factor list
(smoke_factor <- list(no = c('Not at all', 'Rarely'),
                      yes = c('Some days', 'Most days', 'Daily')))

# example data
(x <- data.frame(cp_ptid = paste0("50-22-000", 1:3, "-M1"),
                 en_wrk_cig = c("Some days", "Not at all", "Not at all"),
                 en_wrk_marij = c(rep('Not at all', 3))) )

# assess ETS exposure
wrk_ets_exposed(x, smoke_factor)
}
