\name{abctools-package}
\alias{abctools-package}
\alias{abctools}
\docType{package}
\title{Tools for ABC analyses}
\description{Tools for approximate Bayesian computation including summary statistic selection and assessing coverage}
\details{
\tabular{ll}{
Package: \tab abctools\cr
Type: \tab Package\cr
Version: \tab 0.2-2\cr
Date: \tab 12-12-2013\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{Matt Nunes and Dennis Prangle

Maintainer: Matthew Nunes <m.nunes@lancaster.ac.uk>
}
\references{
For details on methods for summary statistics selection, see

Blum, M. G. B, Nunes, M. A., Prangle, D. and Sisson, S. A. (2013) A
comparative review of dimension reduction methods in approximate
Bayesian computation. \emph{Stat. Sci.} \bold{28}, Issue 2, 189--208.\cr\cr
Fearnhead, P. and Prangle, D. (2012) Constructing summary statistics for approximate Bayesian computation: semi-automatic approximate Bayesian computation. \emph{J. R. Stat. Soc. B} \bold{74}, Part 3, 1--28.\cr\cr
Joyce, P. and P. Marjoram (2008) Approximately sufficient statistics and Bayesian computation. \emph{Stat. Appl. Gen. Mol. Biol.} \bold{7} Article 26.\cr\cr
Nunes, M. A. and Balding, D. J. (2010) On Optimal Selection of Summary Statistics for Approximate Bayesian Computation.  \emph{Stat. Appl. Gen. Mol. Biol.} \bold{9}, Iss. 1, Art. 34. \cr\cr
Wegmann, D. et al. (2009) Efficient approximate Bayesian computation coupled With Markov chain Monte Carlo Without Likelihood. \emph{Genetics} \bold{182} (4), 1207--1218.

For extra scripts used for \code{pls.abc}, see:

Wegmann, D. et al. (2010) ABCtoolbox: A versatile toolkit for
approximate Bayesian computations. \emph{BMC Bioinf.} \bold{11},
116--123.

Details of ABC coverage methods appear in:

Prangle D., Blum M. G. B., Popovic G., Sisson S. A. (2013) Diagnostic
tools of approximate Bayesian computation using the coverage
property. \emph{(preprint)} \url{arxiv.org/abs/1301.3166}
}
\keyword{
  package
  methods
  htest
}
\seealso{
  \code{\link{abc}}
}

