% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_yes_no_vars.R
\name{id_yes_no_vars}
\alias{id_yes_no_vars}
\title{ID "Yes/No" Variables}
\usage{
id_yes_no_vars(x)
}
\arguments{
\item{x}{A dataframe.}
}
\value{
A character vector of variable names.
}
\description{
Identify variables that contain "Yes/yes" and "No/no" values.
}
\examples{
(x <- data.frame(var1 = c('Yes', 'yes', 'no', 'No'), var2 = 1:4, var3 = c('yes', rep(NA, 3))))
id_yes_no_vars(x)
}
